Move pages to a shared directory
This way we can use them in both test suites easily
