+++ cypress_shared/pages/apply/confirmDetails.ts
