+++ cypress_shared/pages/apply/enterCrn.ts
