+++ cypress_shared/pages/apply/index.ts
