+++ cypress_shared/pages/apply/list.ts
