+++ cypress_shared/pages/apply/sentenceType.ts
