+++ cypress_shared/pages/apply/situationPage.ts
