+++ cypress_shared/pages/apply/startPage.ts
