+++ cypress_shared/pages/authManageDetails.ts
