+++ cypress_shared/pages/manage/arrivalCreate.ts
