+++ cypress_shared/pages/manage/booking/confirmation.ts
