+++ cypress_shared/pages/manage/booking/extension/confirmation.ts
