+++ cypress_shared/pages/manage/booking/extension/create.ts
