+++ cypress_shared/pages/manage/booking/new.ts
