+++ cypress_shared/pages/manage/booking/show.ts
