+++ cypress_shared/pages/manage/cancellationConfirmation.ts
