+++ cypress_shared/pages/manage/cancellationCreate.ts
