+++ cypress_shared/pages/manage/departureConfirmation.ts
