+++ cypress_shared/pages/manage/departureCreate.ts
