+++ cypress_shared/pages/manage/index.ts
