+++ cypress_shared/pages/manage/lostBedCreate.ts
